a2a-reminder-agent/agents/reminder/pyproject.toml
[project]
name = "a2a-reminder-agent-module"
version = "0.1.0"
description = "Reminder agent with A2A Protocol built with LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.27.0",
    "langchain>=0.3.25",
    "langchain_openai>=0.0.8",
    "langgraph>=0.4.3",
    "apscheduler>=3.10.4",
    "pydantic>=2.7.4",
    "python-dotenv>=1.0.0",
    "mcp>=1.1.1",
    "mcp-datetime>=0.1.4",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

# No external sources needed

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
